requirements.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
bcrypt==4.0.1
billiard==4.2.0
blinker==1.7.0
cachelib==0.9.0
celery==5.3.5
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
cors==1.0.1
cycler==0.12.1
dnspython==2.4.2
email-validator==2.1.0.post1
filelock==3.16.1
Flask==2.2.3
Flask-Caching==2.1.0
Flask-Cors==5.0.0
Flask-Excel==0.0.7
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.1.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.1
fonttools==4.54.1
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
idna==3.6
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.7
kombu==5.3.4
lml==0.1.0
MarkupSafe==2.1.3
matplotlib==3.9.2
numpy==2.0.2
packaging==24.2
passlib==1.7.4
Pillow==10.1.0
prompt-toolkit==3.0.41
pycparser==2.22
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
PyJWT==2.9.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
reportlab==4.2.5
requests==2.32.3
requests-file==2.1.0
six==1.16.0
SQLAlchemy==2.0.23
texttable==1.7.0
tldextract==5.1.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.12
Werkzeug==2.2.2
WTForms==3.1.0
zipp==3.21.0
zope.event==5.0
zope.interface==7.2